requirements.txt
# Core dependencies for BLE audio transcription
nemo_toolkit[asr]>=2.2.1  # For NVIDIA Parakeet TDT
torch>=2.1.0              # ML backend
numpy<2.0.0               # For NeMo compatibility
soundfile>=0.13.1         # For audio file handling
bleak>=0.21.1             # For BLE communication
opuslib>=3.0.1            # For Opus decoding (requires system libopus)
webrtcvad>=2.0.10         # For voice activity detection
tqdm>=4.67.1              # For progress bars (disabled but required by NeMo)
psycopg2-binary>=2.9.9    # PostgreSQL client
pydantic<2.0.0            # Required by other dependencies
protobuf==3.20.3          # Avoid protobuf 5.x incompat with pydantic v1
sounddevice>=0.4.6        # For microphone capture
python-dotenv>=1.0.0      # For .env loading
